% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr-class.R
\name{as_qtr}
\alias{as_qtr}
\title{Cast an object to qtr}
\usage{
as_qtr(x, origin = c("calendar", "fiscal"))
}
\arguments{
\item{x}{An object}

\item{origin}{Either 'calendar' or 'fiscal'}
}
\value{
An object of class 'qtr'
}
\description{
Cast an object to qtr
}
