% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter-vecs.R
\docType{data}
\name{fy_quarter_starts}
\alias{fy_quarter_starts}
\title{Fiscal quarter start dates}
\format{
## `fy_quarter_starts`
A named vector of length 162.
}
\usage{
fy_quarter_starts
}
\description{
Fiscal quarter start dates
}
\keyword{datasets}
