% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr-class.R
\name{new_qtr}
\alias{new_qtr}
\title{Internal constructor to create qtr type
Can be cast from either character or double}
\usage{
new_qtr(x = double(), origin = c("calendar", "fiscal"))
}
\description{
Internal constructor to create qtr type
Can be cast from either character or double
}
\keyword{internal}
