% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{origins_check}
\alias{origins_check}
\title{Compare the origins of two qtr objects.}
\usage{
origins_check(x, y)
}
\arguments{
\item{x}{an object of class qtr.}

\item{y}{an object of class qtr.}
}
\value{
logical
}
\description{
Compare the origins of two qtr objects.
}
