% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr-class.R
\name{qtr}
\alias{qtr}
\title{A class for quarters}
\usage{
qtr(x, origin = c("calendar", "fiscal"))
}
\arguments{
\item{x}{A character or double vector}

\item{origin}{either 'cs' or 'sims'}
}
\description{
A class for quarters
}
