% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter-vecs.R
\docType{data}
\name{quarter_ends}
\alias{quarter_ends}
\title{Quarter end dates}
\format{
## `quarter_ends`
A named vector of length 164.
}
\usage{
quarter_ends
}
\description{
Quarter end dates
}
\keyword{datasets}
