% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter-vecs.R
\docType{data}
\name{quarter_starts}
\alias{quarter_starts}
\title{Quarter start dates}
\format{
## `quarter_starts`
A named vector of length 164.
}
\usage{
quarter_starts
}
\description{
Quarter start dates
}
\keyword{datasets}
