% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R, R/casting.R, R/coercion.R,
%   R/comparison.R
\name{vec_arith.qtr}
\alias{vec_arith.qtr}
\alias{vec_arith.qtr.default}
\alias{vec_arith.qtr.double}
\alias{vec_arith.qtr.qtr}
\alias{vec_cast.qtr}
\alias{vec_ptype2.qtr}
\alias{vec_proxy_compare.qtr}
\alias{vec_proxy_equal.qtr}
\title{Math stuff for qtr objects}
\usage{
\method{vec_arith}{qtr}(op, x, y)

\method{vec_arith.qtr}{default}(op, x, y)

\method{vec_arith.qtr}{double}(op, x, y)

\method{vec_arith.qtr}{qtr}(op, x, y)

\method{vec_cast}{qtr}(x, to, ...)

\method{vec_ptype2}{qtr}(x, y, ...)

\method{vec_proxy_compare}{qtr}(x, ...)

\method{vec_proxy_equal}{qtr}(x, ...)
}
\arguments{
\item{op}{Arithmetic operation}

\item{x}{qtr object}

\item{y}{integer}
}
\description{
Math stuff for qtr objects
}
