% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qtr_start}
\alias{qtr_start}
\alias{qtr_end}
\alias{qtr_year}
\alias{qtr_qtr}
\alias{qtr_convert}
\title{Utility functions with qtr}
\usage{
qtr_start(x)

qtr_end(x)

qtr_year(x)

qtr_qtr(x)

qtr_convert(x)
}
\arguments{
\item{x}{qtr object}

\item{orders}{character vector of date-time formats, passed to parse_date_time}
}
\value{
character
}
\description{
Utility functions with qtr
}
\examples{
q <- qtr(20204, "calendar")
qtr_start(q)
qtr_end(q)
qtr_year(q)
qtr_qtr(q)
qtr_convert(q)
}
